added api key endpoint and separated the characters logic into a separate routing file
